refactor(auth): migrate authentication logic to a new server module for better organization and maintainability refactor(api): update API routes to use the new authentication module for user session management refactor(project): update project-related functions to utilize the new authentication logic for user validation fix(chat): ensure chat message handling uses the correct user ID from the new authentication module fix(preview): update preview routes to check user access using the new authentication logic fix(files): ensure file access routes validate user permissions with the new authentication module fix(subscription): update subscription status route to use the new authentication logic fix(user): update user-related routes to utilize the new authentication module for session management
