+++ app/api/projects/[id]/download/route.ts
@@ -1,11 +1,11 @@
-import { NextRequest, NextResponse } from 'next/server';
-import { getSession } from '@/lib/auth/session';
-import { getProjectById } from '@/lib/db/projects';
-import { join } from 'path';
-import { exec } from 'child_process';
-import { promisify } from 'util';
-import { readFile, unlink } from 'fs/promises';
+import { auth } from '@/lib/auth/server';
 import { CONTEXT } from '@/lib/constants';
+import { getProjectById } from '@/lib/db/projects';
+import { exec } from 'child_process';
+import { readFile, unlink } from 'fs/promises';
+import { NextRequest, NextResponse } from 'next/server';
+import { join } from 'path';
+import { promisify } from 'util';
 
 const execAsync = promisify(exec);
 
@@ -14,44 +14,44 @@ export async function GET(
   { params }: { params: { id: string } }
 ) {
   try {
-    const session = await getSession();
-    
-    if (!session) {
+    const { userId } = await auth();
+
+    if (!userId) {
       return new NextResponse('Unauthorized', { status: 401 });
     }
-    
+
     const { id } = await Promise.resolve(params);
     const projectId = Number(id);
     if (isNaN(projectId)) {
       return new NextResponse('Invalid project ID', { status: 400 });
     }
-    
+
     const project = await getProjectById(projectId);
-    
-    if (!project || project.createdBy !== session.user.id) {
+
+    if (!project || project.createdBy !== userId) {
       return new NextResponse('Project not found', { status: 404 });
     }
-    
+
     // Get the project directory path
     const projectDir = join(process.cwd(), 'projects', projectId.toString());
     const zipFileName = `${project.name.replace(/[^a-zA-Z0-9]/g, '_')}_${projectId}.zip`;
     const zipFilePath = join('/tmp', zipFileName);
-    
+
     // Create exclude patterns for zip command
     const excludePatterns = [
       ...CONTEXT.EXCLUDE_DIRS.map(dir => `-x "*${dir}/*"`),
       ...CONTEXT.EXCLUDE_FILES.map(file => `-x "${file}"`)
     ].join(' ');
-    
+
     // Create zip file with exclusions
     await execAsync(`cd "${projectDir}" && zip -r "${zipFilePath}" . ${excludePatterns}`);
-    
+
     // Read the zip file using fs.readFile
     const zipBuffer = await readFile(zipFilePath);
-    
+
     // Delete the temporary zip file
     await unlink(zipFilePath);
-    
+
     // Return the zip file
     return new NextResponse(zipBuffer, {
       headers: {
@@ -62,8 +62,8 @@ export async function GET(
   } catch (error) {
     console.error('Error downloading project:', error);
     return new NextResponse(
-      JSON.stringify({ error: error instanceof Error ? error.message : 'Internal Server Error' }), 
-      { 
+      JSON.stringify({ error: error instanceof Error ? error.message : 'Internal Server Error' }),
+      {
         status: 500,
         headers: {
           'Content-Type': 'application/json',
@@ -71,4 +71,4 @@ export async function GET(
       }
     );
   }
-} 
+}
