+++ app/api/projects/[id]/preview/route.ts
@@ -1,4 +1,4 @@
-import { getSession } from '@/lib/auth/session';
+import { auth } from '@/lib/auth/server';
 import { getProjectById } from '@/lib/db/projects';
 import { NextRequest, NextResponse } from 'next/server';
 
@@ -12,8 +12,8 @@ export async function GET(
 ) {
   try {
     // Get the session
-    const session = await getSession();
-    if (!session) {
+    const { userId } = await auth();
+    if (!userId) {
       return NextResponse.json(
         { error: 'Unauthorized' },
         { status: 401 }
@@ -42,8 +42,8 @@ export async function GET(
     }
 
     // Check if the user has access to the project
-    if (project.createdBy !== session.user.id) {
-      console.log(`[Preview API] User ${session.user.id} does not have access to project ${projectId}`);
+    if (project.createdBy !== userId) {
+      console.log(`[Preview API] User ${userId} does not have access to project ${projectId}`);
       return NextResponse.json(
         { error: 'Forbidden' },
         { status: 403 }
@@ -106,8 +106,8 @@ export async function POST(
 ) {
   try {
     // Get the session
-    const session = await getSession();
-    if (!session) {
+    const { userId } = await auth();
+    if (!userId) {
       return NextResponse.json(
         { error: 'Unauthorized' },
         { status: 401 }
@@ -133,7 +133,7 @@ export async function POST(
     }
 
     // Check if the user has access to the project
-    if (project.createdBy !== session.user.id) {
+    if (project.createdBy !== userId) {
       return NextResponse.json(
         { error: 'Forbidden' },
         { status: 403 }
@@ -183,8 +183,8 @@ export async function DELETE(
 ) {
   try {
     // Get the session
-    const session = await getSession();
-    if (!session) {
+    const { userId } = await auth();
+    if (!userId) {
       return NextResponse.json(
         { error: 'Unauthorized' },
         { status: 401 }
@@ -210,7 +210,7 @@ export async function DELETE(
     }
 
     // Check if the user has access to the project
-    if (project.createdBy !== session.user.id) {
+    if (project.createdBy !== userId) {
       return NextResponse.json(
         { error: 'Forbidden' },
         { status: 403 }
