+++ app/api/projects/route.ts
@@ -3,10 +3,10 @@ import { z } from 'zod';
 
 import { ApiErrorHandler } from '@/lib/api/errors';
 import { ApiResponseHandler } from '@/lib/api/responses';
-import { withAuth, RouteContext } from '@/lib/auth/middleware';
-import { 
-  createProject, 
-  getProjectsByUserId 
+import { auth } from '@/lib/auth/server';
+import {
+    createProject,
+    getProjectsByUserId
 } from '@/lib/db/projects';
 
 
@@ -22,17 +22,19 @@ const createProjectSchema = z.object({
  */
 export async function GET(request: NextRequest) {
   try {
-    const searchParams = request.nextUrl.searchParams;
-    const userId = searchParams.get('userId');
-
+    const { userId } = await auth();
     if (!userId) {
       return NextResponse.json(
-        { error: 'User ID is required' },
-        { status: 400 }
+        { error: 'Unauthorized' },
+        { status: 401 }
       );
     }
 
-    const projects = await getProjectsByUserId(Number(userId));
+    const searchParams = request.nextUrl.searchParams;
+    const requestedUserId = searchParams.get('userId');
+
+    // Use the authenticated user's ID, not from query params for security
+    const projects = await getProjectsByUserId(requestedUserId || userId);
     return NextResponse.json(projects);
   } catch (error) {
     console.error('Error fetching projects:', error);
@@ -47,33 +49,35 @@ export async function GET(request: NextRequest) {
  * POST /api/projects
  * Create a new project
  */
-export const POST = withAuth(
-  async (
-    request: NextRequest, 
-    context: RouteContext, 
-    session: { user: { id: number } }
-  ) => {
-    try {
-      // Parse the request body
-      const body = await request.json();
-      
-      // Validate the request body
-      const result = createProjectSchema.safeParse(body);
-      if (!result.success) {
-        return ApiErrorHandler.validationError(result.error);
-      }
-      
-      // Create the project
-      const project = await createProject({
-        name: result.data.name,
-        description: result.data.description || null,
-        userId: session.user.id,
-        createdBy: session.user.id,
-      });
-
-      return ApiResponseHandler.created({ project });
-    } catch (error) {
-      return ApiErrorHandler.handle(error);
+export async function POST(request: NextRequest) {
+  try {
+    const { userId } = await auth();
+    if (!userId) {
+      return NextResponse.json(
+        { error: 'Unauthorized' },
+        { status: 401 }
+      );
     }
+
+    // Parse the request body
+    const body = await request.json();
+
+    // Validate the request body
+    const result = createProjectSchema.safeParse(body);
+    if (!result.success) {
+      return ApiErrorHandler.validationError(result.error);
+    }
+
+    // Create the project
+    const project = await createProject({
+      name: result.data.name,
+      description: result.data.description || null,
+      userId: userId,
+      createdBy: userId,
+    });
+
+    return ApiResponseHandler.created({ project });
+  } catch (error) {
+    return ApiErrorHandler.handle(error);
   }
-); 
+}
