+++ app/api/subscription/status/route.ts
@@ -1,4 +1,4 @@
-import { auth } from '@/lib/auth';
+import { auth } from '@/lib/auth/server';
 import { NextResponse } from 'next/server';
 
 export async function GET() {
