+++ app/api/user/delete-account/route.ts
@@ -1,4 +1,4 @@
-import { auth } from '@/lib/auth';
+import { auth } from '@/lib/auth/server';
 import { db } from '@/lib/db/drizzle';
 import { users } from '@/lib/db/schema';
 import { createClerkClient } from '@clerk/nextjs/server';
